move schemaProps and custom function
