+++ lib/header.js
@@ -7,6 +7,28 @@
 
 var _ = require('lodash');
 
+function custom(schema) {
+  if (schema.allOf) {
+    for (let i=0; i<schema.allOf.length; i++) {
+      if (schema.allOf[i].$ref && schema.allOf[i].$ref === 'https://ns.adobe.com/xdm/common/extensible.schema.json#/definitions/@context') {
+        return true;
+      }
+    }
+  }
+  return false;
+}
+
+function schemaProps(schema, schemaPath, filename) {
+  return {
+    // if there are definitions, but no properties
+    abstract: (schema.definitions !== undefined && _.keys(schema.properties).length === 0) ? 'Cannot be instantiated' : 'Can be instantiated',
+    extensible: (schema.definitions !== undefined || schema['meta:extensible'] === true) ? 'Yes' : 'No',
+    status: schema['meta:status'] !== undefined ? (schema['meta:status'].charAt(0).toUpperCase() + schema['meta:status'].slice(1)) : 'Experimental',
+    custom: custom(schema) ? 'Allowed' : 'Forbidden',
+    original: filename.substr(schemaPath.length).substr(1),
+  };
+}
+
 class Header {
   constructor(name, docs, value, links) {
     this.name = name;
@@ -51,7 +73,8 @@ function header(name, docs, value, links) {
   return this;
 }
 
-function headers(schema, props, docs) {
+function headers(schema, path, filename, docs) {
+  const props = schemaProps(schema, path, filename);
   this.doclinks = docs ? docs : {};
   this.myheaders = [];
