+++ lib/markdownWriter.js
@@ -30,28 +30,6 @@ function assoc(obj, key, value) {
   return obj;
 }
 
-function custom(schema) {
-  if (schema.allOf) {
-    for (let i=0; i<schema.allOf.length; i++) {
-      if (schema.allOf[i].$ref && schema.allOf[i].$ref === 'https://ns.adobe.com/xdm/common/extensible.schema.json#/definitions/@context') {
-        return true;
-      }
-    }
-  }
-  return false;
-}
-
-function schemaProps(schema, schemaPath, filename) {
-  return {
-    // if there are definitions, but no properties
-    abstract: (schema.definitions !== undefined && _.keys(schema.properties).length === 0) ? 'Cannot be instantiated' : 'Can be instantiated',
-    extensible: (schema.definitions !== undefined || schema['meta:extensible'] === true) ? 'Yes' : 'No',
-    status: schema['meta:status'] !== undefined ? (schema['meta:status'].charAt(0).toUpperCase() + schema['meta:status'].slice(1)) : 'Experimental',
-    custom: custom(schema) ? 'Allowed' : 'Forbidden',
-    original: filename.substr(schemaPath.length).substr(1),
-  };
-}
-
 function flatten(dependencies) {
   let deps = [];
   if (dependencies) {
@@ -173,8 +151,7 @@ const generateMarkdown = function(filename, schema, schemaPath, outDir, dependen
     [ 'header.ejs', {
       schema: schema,
       dependencies: flatten(dependencyMap),
-      table: headers(schema, schemaProps(schema, schemaPath, filename)).render(),
-      props: schemaProps(schema, schemaPath, filename) } ],
+      table: headers(schema, schemaPath, filename).render() } ],
     //[ 'divider.ejs', null ],
     //[ 'topSchema.ejs', ctx ],
     [ 'examples.ejs', { examples: stringifyExamples(schema.examples), title: schema.title } ]
