+++ package.json
@@ -30,6 +30,7 @@
     "eslint": "^4.10.0",
     "istanbul": "^0.4.5",
     "jasmine": "~2.8.0",
+    "jasmine-diff": "^0.1.3",
     "npm-snapshot": "^1.0.3"
   },
   "engines": {
